+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## unreleased
 * BraintreePayPal
   * Fix an issue where `BTPayPalRequest` was sending `phone_number` instead of `payer_phone`
+  * Add `merchant` and `flow_type` as query parameters to the app switch URL.
 
 ## 6.36.0 (2025-08-13)
 * BraintreeCore